+++ tests/Console/Command/CacheOpenAPIRoutesTest.php
@@ -7,9 +7,11 @@ namespace Membrane\OpenAPIRouter\Tests\Console\Command;
 use Membrane\OpenAPIRouter\Console\Command\CacheOpenAPIRoutes;
 use Membrane\OpenAPIRouter\Console\Service;
 use Membrane\OpenAPIRouter\Exception;
-use Membrane\OpenAPIRouter\Router\Collector\RouteCollector;
+use Membrane\OpenAPIRouter\Route\Path;
+use Membrane\OpenAPIRouter\Route\Server;
+use Membrane\OpenAPIRouter\RouteCollection;
+use Membrane\OpenAPIRouter\RouteCollector;
 use Membrane\OpenAPIRouter\Router\Route;
-use Membrane\OpenAPIRouter\Router\RouteCollection;
 use org\bovigo\vfs\vfsStream;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -22,7 +24,7 @@ use Symfony\Component\Console\Tester\CommandTester;
 #[CoversClass(CacheOpenAPIRoutes::class)]
 #[CoversClass(Exception\CannotCollectRoutes::class)]
 #[UsesClass(Service\CacheOpenAPIRoutes::class)]
-#[UsesClass(Route\Route::class), UsesClass(Route\Server::class), UsesClass(Route\Path::class)]
+#[UsesClass(\Membrane\OpenAPIRouter\Route\Route::class), UsesClass(Server::class), UsesClass(Path::class)]
 #[UsesClass(RouteCollection::class)]
 #[UsesClass(RouteCollector::class)]
 class CacheOpenAPIRoutesTest extends TestCase
